Implemented `useColumnUpdates` hook
